Float type case fixing

+++ Stage.py
@@ -201,11 +201,11 @@ class Stage:
                 line = file.readline()
                 if "dur=" not in line or not line or '\n' not in line:
                     break
-                targetData.append(int(line[4:-1]))
+                targetData.append(float(line[4:-1]))
                 line = file.readline()
                 if "del=" not in line or not line or '\n' not in line:
                     break
-                targetData.append(int(line[4:-1]))
+                targetData.append(float(line[4:-1]))
                 line = file.readline()
                 if "val=" not in line or not line or '\n' not in line:
                     break
